Add psuedo r-squred values
This adds support for calculating psuedo-r2 values. To do so, functions for loglikelihood and nullloglikelihood are added and the output type needs to store a few extra variables (y, mu and a dof).

One additional change I considered but did not implement is that the biascorrection tests often need the y and/or mu values, expecting these to be in the augment_df argument. Saving them all the time in order to calculate pseudo-r2 might make it unnecessary to keep those in the augment_df